Created launch.json file for default config on VSCode debugging feature.
Previously, in order to make the instance of the running app available to all devices on the network by accessing the host IP and port on the browser, an also use a specific web renderer (html, in our case), we had to run the following command: $ flutter run -d web-server --web-hostname 0.0.0.0 --web-port 8080. But, when using $ flutter run to run the app, we dont have access to automatic Hot Reload.

On the other hand, when using VSCode Debug Feature (Run -> Start Debugging), we have access to automatic Hot Reload. Therefore, in order to have both the arguments set and also have access to Hot Reload, we created the launch.json file, where we can set default arguments to be used when running VSCode Debug.